+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -   The `string.template` node now supports passing object keys as variables for string expansion.
 
+-   Added the `input.object.filter-duplicates` and `input.object.group` nodes. These allow duplication filtering and grouping based on an object key.
+
 -   Added the `string.join` node, which allows you to join multiple strings together. A seperator can be introduced between the strings.
 
 -   Support for `script type="module"` in node editor contents. Note that these script elements are untouched when they are processed by the browser. Script elements that do not contain the `type="module"` directive will be evaluated like tever before, with the document argument attached to the shadow root and `this` referring to the node itself.